Fix display id to long, news sample

+++ src/site/display.c
@@ -20,14 +20,14 @@ int main () {
 	cgi_out = stdout;
 
 	char * data = getenv("QUERY_STRING");
-	int id;
+	long id;
 	if(data == NULL){
 		// TODO: add error message out macro/function to decorator, EG 404
 		cgiHeader(NULL);
 		cgiOut("<P>Error! Error in passing data from form to script.");
 	    cgiFooter(NULL);
 	}
-	else if(sscanf(data,"id=%d",&id)!=1){
+	else if(sscanf(data,"id=%ld",&id)!=1){
 		cgiHeader(NULL);
 		cgiOut("<P>Error! Invalid data. Data must be numeric.");
 		cgiFooter(NULL);
@@ -38,7 +38,7 @@ int main () {
 		// TODO: fix return value evaluations and exceptions
 		if (NULL == path){
 			cgiHeader(NULL);
-			cgiOut("<P>Error! Path data. %d</p>", id);
+			cgiOut("<P>Error! Path data. %ld</p>", id);
 		    cgiFooter(NULL);
 		    return 0;
 		}
